% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindGroupDEGs.R
\name{FindGroupDEGs}
\alias{FindGroupDEGs}
\title{FindGroupDEGs}
\usage{
FindGroupDEGs(object, ident_use, compare_by, test_use = "wilcox",
  min_fold_change = 0, min_pct_express_diff = 0, pval_thresh = 0.05,
  genes_of_interest = NULL, cell_number_thresh = 10,
  pval_use = p_val_adj)
}
\arguments{
\item{object}{Processed Seurat scRNAseq object}

\item{ident_use}{Identity (from the ident slot or a meta.data column) by
which to group the cells.}

\item{compare_by}{Meta.data column by which to group the cells within each
identity for comparison.}

\item{test_use}{Differential expression test to use.  Same values as in
FindMarkers/FindAllMarkers (i.e wilcox, MAST, DESeq2, ...). Default:
'wilcox'}

\item{min_fold_change}{Minimum log fold change difference between two groups}

\item{min_pct_express_diff}{Minimum difference in the percentage of each group
that expresses the gene in question above a threshold}

\item{pval_thresh}{Signifigance cutoff for DE testing. Default: 0.05}

\item{genes_of_interest}{A list of genes to use in DE testing. Default: NULL}

\item{cell_number_thresh}{Minimum number of cells that must be within an
ident.use/group.by grouping before it is discarded. Default: 10}

\item{pval_use}{p value to use when filtering DE genes by the pval.thresh.
Acceptable values are p_val_adj and p_val.  Default: p_val_adj}
}
\value{
Named list with genes
}
\description{
Find differentially expressed genes within an identity
  contrasted by a grouping variable (e.g. comparison between disease status
  or drug treatment (i.e. compare_by) within each cell type (i.e. ident_use).
  Performs 1-to-1 comparisons for each possible of combinations of the values
  of the grouping variable.
}
