% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDEGenes.R
\name{PlotDEGenes}
\alias{PlotDEGenes}
\title{PlotDEGenes}
\usage{
PlotDEGenes(object, combo_list, group_by = "celltype",
  compare_by = "class", gene_annotation_df = NULL, split_if = NULL,
  ...)
}
\arguments{
\item{object}{Processed Seurat scRNAseq object}

\item{combo_list}{Differentially expressed gene list for combination of
contrast factors from FindGroupDEGs}

\item{group_by}{Group variable within which to examine the cells.
Default: celltype}

\item{compare_by}{Group variable to use when comparing the cells.
Default: class.}

\item{...}{Extra arguments to pass to seuratBubblePlot::bubbleplot()}
}
\value{
ggplot2 object
}
\description{
Using a bubbleplot, plot genes that are differentially expressed between
cells of a comparison group within a grouping (i.e. DEGs between different
disease groups within the same cell type).  The results of all intragroup
comparisons are compiled together and plotted together on one bubbleplot.
}
