% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSplitGroupDEGs.R
\name{PlotSplitGroupDEGs}
\alias{PlotSplitGroupDEGs}
\title{PlotSplitGroupDEGs}
\usage{
PlotSplitGroupDEGs(seuratObj, combo_list, group_by = "celltype",
  compare_by = "class", ...)
}
\arguments{
\item{seuratObj}{Processed Seurat scRNAseq object}

\item{combo_list}{List of list of differentially expressed genes from DEcombinations}

\item{group_by}{Primary grouping variable (from ident slot or meta.data column)}

\item{compare_by}{Intragroup comparison variable}

\item{...}{Additional parameters to pass to bubbleplot}
}
\value{
A list of gtables
}
\description{
For a given Seurat object and a list of intragroup DEGs, for each grouping
plot the genes that were differentially expressed between each comparison_group.
The results for each comparison within the group will be plotted in its own
bubbleplot, and all comparisons for that group are combined into one compound plot.
}
