% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveDEGcomparePlots.R
\name{saveDEGcomparePlots}
\alias{saveDEGcomparePlots}
\title{saveDEGcomparePlots}
\usage{
saveDEGcomparePlots(plot_list, path = "~/", format = "png")
}
\arguments{
\item{plot_list}{List of group DEGs as generated by PlotDEGenes}

\item{path}{Location in which to save the plots. Default = ~/}

\item{format}{Filetype to save the file to.  Save options as ggsave,
that is "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png",
"bmp", "svg" or "wmf".  Default = png}
}
\description{
Save each of the plots generated by PlotDEGenes to its own
  appropriately scaled file.
}
