% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteDECombinations.R
\name{WriteFindGroupDEGs}
\alias{WriteFindGroupDEGs}
\title{WriteGroupDEGs}
\usage{
WriteFindGroupDEGs(combo_list, location)
}
\arguments{
\item{combo_list}{Differentially expressed gene list for combination of contrast factors from FindGroupDEGs}

\item{location}{Location to which to write the csv files}
}
\description{
Given a list from FindGroupDEGs, write the expression values and percentage
of cells expressing to csv files
}
